@@ -24,18 +24,18 @@
   "dependencies": {
     "nock": "^10.0.6",
     "@types/nock": "^10.0.3",
-    "probot": "^9.2.11"
+    "probot": "^9.2.19"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.13",
-    "@types/node": "^12.0.2",
+    "@types/jest": "^24.0.15",
+    "@types/node": "^12.6.2",
     "eslint-plugin-typescript": "^0.14.0",
     "jest": "^24.8.0",
-    "nodemon": "^1.19.0",
+    "nodemon": "^1.19.1",
     "smee-client": "^1.1.0",
-    "standard": "^12.0.1",
+    "standard": "^13.0.2",
     "ts-jest": "^24.0.2",
-    "typescript": "^3.4.5",
+    "typescript": "^3.5.3",
     "typescript-eslint-parser": "^22.0.0"
   },
   "engines": {
