@@ -46,7 +46,7 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
     }
 };
 var _this = this;
-var _ = require('lodash');
+var _ = require('lodash'); // eslint-disable-line no-unused-vars
 var ProbotRequest = /** @class */ (function () {
     function ProbotRequest(robot, context) {
         this.isCollaborator = false;
@@ -62,7 +62,7 @@ var ProbotRequest = /** @class */ (function () {
             feedback: '',
             no_debug_id: '',
             ignore: [],
-            reopen: [],
+            reopen: []
         };
     }
     ProbotRequest.prototype.load = function () {
@@ -109,9 +109,9 @@ var ProbotRequest = /** @class */ (function () {
                         this.labels = this.issue.labels.slice();
                         this.isBot = this.context.payload.sender.login.endsWith('[bot]');
                         this.ignore = (_.intersection(this.issue.labels, this.config.ignore).length !== 0) || this.isBot;
-                        this.reopen = (_.intersection(this.issue.labels.concat('*'), this.config.reopen).length !== 0)
-                            && (_.intersection(this.issue.labels.map(function (label) { return "-" + label; }), this.config.reopen).length === 0)
-                            && !this.isBot;
+                        this.reopen = (_.intersection(this.issue.labels.concat('*'), this.config.reopen).length !== 0) &&
+                            (_.intersection(this.issue.labels.map(function (label) { return "-" + label; }), this.config.reopen).length === 0) &&
+                            !this.isBot;
                         return [2 /*return*/, this];
                 }
             });
@@ -162,17 +162,20 @@ module.exports = function (robot) {
                     req = _a.sent();
                     if (req.isBot || req.isCollaborator)
                         return [2 /*return*/];
-                    if (!(!req.labels.includes('question') && !req.body.match(/\b[A-Z0-9]{8}-(euc|apse)\b/))) return [3 /*break*/, 4];
+                    if (!req.config.no_debug_id)
+                        return [2 /*return*/];
+                    if (req.labels.includes('question'))
+                        return [2 /*return*/];
+                    if (req.labels.includes('enhancement'))
+                        return [2 /*return*/];
+                    if (req.body.match(/\b[A-Z0-9]{8}-(euc|apse)\b/))
+                        return [2 /*return*/];
                     req.label(req.config.feedback);
                     return [4 /*yield*/, req.save('issues.opened')];
                 case 2:
                     _a.sent();
-                    if (!req.config.no_debug_id) return [3 /*break*/, 4];
-                    return [4 /*yield*/, context.github.issues.createComment(context.issue({ body: req.config.no_debug_id }))];
-                case 3:
-                    _a.sent();
-                    _a.label = 4;
-                case 4: return [2 /*return*/];
+                    context.github.issues.createComment(context.issue({ body: req.config.no_debug_id }));
+                    return [2 /*return*/];
             }
         });
     }); });
